polish(mcp): set error_kind=scope_denied; tighten DeleteTagConflict type
Address three code-review findings on the whole-diff pass:

1. `scope_denied_error` now sets `error_kind = "scope_denied"`, matching
   the programmatic contract documented in the new mcp-modes.mdx
   decision tree and aligning with other typed error_kinds emitted by
   the MCP surface (`missing_parent_workspace_id`,
   `missing_project_id`, `session_busy`, …). Locked in with a new unit
   test in `response_classification`.

2. `check_scope_allows_workspace`'s doc comment was stale — it still
   referred to "orchestrator mode" and the "non-orchestrator mode"
   fallback, but the body was updated for Workspace mode in commit
   c7b43ab9. Rewritten to describe Global / scope-missing / self /
   child / denied, explicitly naming both Workspace and Orchestrator.

3. `DeleteTagConflict.issue_tag_count` changed from `i64` to `i32`.
   ts-rs maps `i64` → `bigint`, which stock `JSON.parse` can't produce
   without a reviver — every frontend consumer would have had to
   handle a BigInt for a field that never realistically exceeds
   `i32::MAX`. Backend clamps via `i32::try_from(count).unwrap_or(i32::MAX)`
   and `shared/types.ts` now reads `number`.

+++ crates/api-types/src/tag.rs
@@ -44,5 +44,5 @@ pub struct ListTagsResponse {
 #[derive(Debug, Serialize, TS)]
 pub struct DeleteTagConflict {
     pub message: String,
-    pub issue_tag_count: i64,
+    pub issue_tag_count: i32,
 }
